# perf: complete TASK-061 through TASK-065 - comprehensive performance testing and optimizations