--- a/TASKS.md
+++ b/TASKS.md
@@ -125,11 +125,11 @@
 - [ ] **TASK-060**: Test thread safety and concurrency
 
 ### Performance Testing
-- [ ] **TASK-061**: Benchmark Result creation and operations
-- [ ] **TASK-062**: Measure validation performance
-- [ ] **TASK-063**: Profile memory allocations
-- [ ] **TASK-064**: Compare with exception performance
-- [ ] **TASK-065**: Optimize hot paths based on results
+- [x] **TASK-061**: Benchmark Result creation and operations
+- [x] **TASK-062**: Measure validation performance
+- [x] **TASK-063**: Profile memory allocations
+- [x] **TASK-064**: Compare with exception performance
+- [x] **TASK-065**: Optimize hot paths based on results
 
 ### Documentation
 - [x] **TASK-066**: Write comprehensive XML documentation