--- a/src/Core/Results/Result.cs
+++ b/src/Core/Results/Result.cs
@@ -1,4 +1,5 @@
 ﻿using System.Diagnostics.CodeAnalysis;
+using System.Runtime.CompilerServices;
 using System.Security;
 using System.Text.Json.Serialization;
 
@@ -315,10 +316,13 @@ public partial class Result : IResult
     /// Console.WriteLine(message);
     /// </code>
     /// </example>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public TResult Match<TResult>(Func<TResult> onSuccess, Func<string, TResult> onFailure)
     {
+#if DEBUG
         ArgumentNullException.ThrowIfNull(onSuccess);
         ArgumentNullException.ThrowIfNull(onFailure);
+#endif
 
         return IsSuccess
             ? onSuccess()
@@ -612,6 +616,11 @@ public partial class Result : IResult
     public string Error { get; private set; } = string.Empty;
 
     /// <summary>
+    /// Shared empty dictionary to avoid allocations in success cases.
+    /// </summary>
+    private static readonly IDictionary<string, string[]> EmptyFailures = new Dictionary<string, string[]>();
+    
+    /// <summary>
     /// Gets a dictionary of field-specific validation failures.
     /// </summary>
     /// <value>
@@ -624,7 +633,7 @@ public partial class Result : IResult
     /// or properties. For non-validation failures, this dictionary is empty.
     /// </remarks>
     [JsonInclude]
-    public IDictionary<string, string[]> Failures { get; private set; } = new Dictionary<string, string[]>();
+    public IDictionary<string, string[]> Failures { get; private set; } = EmptyFailures;
 
     /// <summary>
     /// Gets the specific type of failure that occurred.
@@ -665,7 +674,7 @@ public partial class Result : IResult
     /// It is always the logical inverse of <see cref="IsFailure"/>.
     /// </remarks>
     public bool IsSuccess =>
-        !IsFailure;
+        string.IsNullOrEmpty(Error);
 
     /// <summary>
     /// Gets the general category of this result.